# go-core/docs/swagger.yaml
basePath: /
definitions:
  data.MarketType:
    enum:
    - indian
    - us
    - crypto
    - forex
    - commodities
    type: string
    x-enum-varnames:
    - MarketTypeIndian
    - MarketTypeUS
    - MarketTypeCrypto
    - MarketTypeForex
    - MarketTypeCommodities
...
  data.OutcomeSummary:
    enum:
    - profitable
    - loss
    - breakeven
    - partial_profit
    - partial_loss
    type: string
    x-enum-varnames:
    - OutcomeSummaryProfitable
    - OutcomeSummaryLoss
    - OutcomeSummaryBreakeven
    - OutcomeSummaryPartialProfit
    - OutcomeSummaryPartialLoss
...
  data.TradeDirection:
    enum:
    - long
    - short
    type: string
    x-enum-varnames:
    - TradeDirectionLong
    - TradeDirectionShort
...
  dto.CreatePsychologyRequest:
    properties:
      emotional_state:
        type: string
      entry_confidence:
        maximum: 10
        minimum: 1
        type: integer
      lessons_learned:
        type: string
      mistakes_made:
        items:
          type: string
        type: array
      satisfaction_rating:
        maximum: 10
        minimum: 1
        type: integer
    required:
    - emotional_state
    - entry_confidence
    - satisfaction_rating
    type: object
...
  dto.CreateTradeRequest:
    properties:
      direction:
        $ref: '#/definitions/data.TradeDirection'
      entry_date:
        type: string
      entry_price:
        type: number
      exit_price:
        type: number
      market_type:
        $ref: '#/definitions/data.MarketType'
      outcome_summary:
        $ref: '#/definitions/data.OutcomeSummary'
      psychology:
        $ref: '#/definitions/dto.CreatePsychologyRequest'
      quantity:
        type: integer
      rules_followed:
        items:
          type: string
        type: array
      screenshots:
        items:
          type: string
        type: array
      stop_loss:
        type: number
      strategy:
        type: string
      symbol:
        type: string
      target:
        type: number
      total_amount:
        type: number
      trade_analysis:
        type: string
      user_id:
        type: integer
    required:
    - direction
    - entry_date
    - entry_price
    - market_type
    - outcome_summary
    - quantity
    - strategy
    - symbol
    - total_amount
    - user_id
    type: object
...
  dto.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      message:
        type: string
    type: object
...
  dto.GetTradesResponse:
    properties:
      pagination:
        $ref: '#/definitions/dto.PaginationResponse'
      trades:
        items:
          $ref: '#/definitions/dto.TradeResponse'
        type: array
    type: object
...
  dto.PaginationResponse:
    properties:
      count:
        type: integer
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  dto.PsychologyResponse:
    properties:
      emotional_state:
        type: string
      entry_confidence:
        type: integer
      lessons_learned:
        type: string
      mistakes_made:
        items:
          type: string
        type: array
      satisfaction_rating:
        type: integer
    type: object
...
  dto.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
  dto.TradeResponse:
    properties:
      created_at:
        type: string
      direction:
        $ref: '#/definitions/data.TradeDirection'
      entry_date:
        type: string
      entry_price:
        type: number
      exit_price:
        type: number
      id:
        type: string
      market_type:
        $ref: '#/definitions/data.MarketType'
      outcome_summary:
        $ref: '#/definitions/data.OutcomeSummary'
      psychology:
        $ref: '#/definitions/dto.PsychologyResponse'
      quantity:
        type: integer
      rules_followed:
        items:
          type: string
        type: array
      screenshots:
        items:
          type: string
        type: array
      stop_loss:
        type: number
      strategy:
        type: string
      symbol:
        type: string
      target:
        type: number
      total_amount:
        type: number
      trade_analysis:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  dto.UpdateTradeRequest:
    properties:
      direction:
        $ref: '#/definitions/data.TradeDirection'
      entry_date:
        type: string
      entry_price:
        type: number
      exit_price:
        type: number
      id:
        type: string
      market_type:
        $ref: '#/definitions/data.MarketType'
      outcome_summary:
        $ref: '#/definitions/data.OutcomeSummary'
      psychology:
        $ref: '#/definitions/dto.CreatePsychologyRequest'
      quantity:
        type: integer
      rules_followed:
        items:
          type: string
        type: array
      screenshots:
        items:
          type: string
        type: array
      stop_loss:
        type: number
      strategy:
        type: string
      symbol:
        type: string
      target:
        type: number
      total_amount:
        type: number
      trade_analysis:
        type: string
      user_id:
        type: integer
    required:
    - direction
    - entry_date
    - entry_price
    - id
    - market_type
    - outcome_summary
    - quantity
    - strategy
    - symbol
    - total_amount
    - user_id
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive trading journal API for tracking trades, setups, and
    market analysis
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: 100xTrader API
  version: "1.0"
paths:
...
  /api/v1/trades:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of trades
      parameters:
      - description: 'Number of trades to return (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      - description: 'Number of trades to skip (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Trades retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.GetTradesResponse'
              type: object
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: List trades
      tags:
      - trades
    post:
      consumes:
      - application/json
      description: Create a new trade with all required and optional fields
      parameters:
      - description: Trade data
        in: body
        name: trade
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTradeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Trade created successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.TradeResponse'
              type: object
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create a new trade
      tags:
      - trades
  /api/v1/trades/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a specific trade by ID
      parameters:
      - description: Trade ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Trade deleted successfully
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Invalid trade ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Trade not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete a trade
      tags:
      - trades
    get:
      consumes:
      - application/json
      description: Retrieve a specific trade with all its details
      parameters:
      - description: Trade ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Trade retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.TradeResponse'
              type: object
        "400":
          description: Invalid trade ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Trade not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get a trade by ID
      tags:
      - trades
    put:
      consumes:
      - application/json
      description: Update an existing trade with new data
      parameters:
      - description: Trade ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated trade data
        in: body
        name: trade
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTradeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Trade updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.TradeResponse'
              type: object
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Trade not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update a trade
      tags:
      - trades
  /api/v1/trades/user/{user_id}:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of trades for a specific user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: 'Number of trades to return (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      - description: 'Number of trades to skip (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User trades retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.GetTradesResponse'
              type: object
        "400":
          description: Invalid user ID or query parameters
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get trades by user
      tags:
      - trades
...
schemes:
- http
swagger: "2.0"